refactor: clean up migration scripts and improve user DTOs
- Simplified SQL statements in migrations for better readability.
- Added a new migration to alter the "monthly_limit" column type in the "user" table.
- Enhanced validation rules in CreateUserDto and UpdateUserDto for password requirements.
- Changed SetLimitDto to use IsNumber instead of IsInt for limit validation.
- Updated UserController and UserService to improve type definitions and readability.
- Refactored CreateUserDialog and ProfileSection components for better code organization and clarity.
- Added admin badge to UserTable for better user role visibility.

+++ backend/src/entities/User.ts
@@ -23,7 +23,7 @@ export class User extends Base {
   @Property({ type: 'boolean', default: false })
   admin: boolean = false;
 
-  @Property({ type: 'integer', nullable: true, fieldName: 'monthly_limit' })
+  @Property({ type: 'float', nullable: true, fieldName: 'monthly_limit' })
   monthlyLimit: number | null = null;
 
   @OneToMany(() => Chat, (chat) => chat.user)
